refactor: feat: split up program into several cs files, resolve TODOs, and change exporting scheme
